Update weatherbot.py
I added: API Token for weather API using requests library. Added a group of greetings to be change or modified and the bot randomized decides which one to choose once is launched. It also recognize the user_id and refers to it. It expect to receive the command !weather {city}, which I would like to get automatic from the system without adding the city by hand. The bot now answers the weather info (condition, temperature, humidity and wind speed. The code is commented.